Move flatdata-py to flatdata.lib
