--- flatdata-py/flatdata/generator
@@ -1,1 +1,0 @@
-../../flatdata-generator/flatdata/generator
