+++ flatdata-py/flatdata/lib/errors.py
