+++ flatdata-py/flatdata/lib/inspector.py
