+++ flatdata-py/flatdata/lib/structure.py
