+++ flatdata-py/setup.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from setuptools import find_packages
+from setuptools import find_namespace_packages
 import setuptools
 
 
@@ -11,7 +11,7 @@ setuptools.setup(
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
     url="https://github.com/heremaps/flatdata",
-    packages=['flatdata'],
+    packages=find_namespace_packages(),
     entry_points={
         'console_scripts': [
             'flatdata-inspector=flatdata.inspector:main'
